+++ webui/components/jobs/detail/DetailTable.vue
