+++ webui/components/jobs/detail/StageSelect.vue
